--- commons/src/main/kotlin/be/solidlab/sdx/client/commons/auth/SolidClientCredentials.kt
@@ -1,7 +1,0 @@
-package be.solidlab.sdx.client.commons.auth
-
-data class SolidClientCredentials(
-    val identityServerUrl: String,
-    val clientId: String,
-    val clientSecret: String
-)
