+++ commons/src/main/kotlin/be/ugent/solidlab/sdx/client/commons/auth/SolidClientCredentials.kt
@@ -1,0 +1,7 @@
+package be.ugent.solidlab.sdx.client.commons.auth
+
+data class SolidClientCredentials(
+    val identityServerUrl: String,
+    val clientId: String,
+    val clientSecret: String
+)
