+++ gradle-plugin/gradle.properties
@@ -1,0 +1,3 @@
+group=be.ugent.solidlab
+version=0.0.1
+name=sdx-plugin
